Fix the automated regression check
It seems that this was a thing before, but I guess it had bit-rotted.
This updates it to use the new `--presubmit` option to the touch tests,
which only fails if a regression occurs in a known test or an unknown
(new) test fails.

BUG=None
TEST=cause a regression and check the presubmit fails

Cq-Depend: chromium:2171940
Change-Id: I6903056d2b180c6a0fd607f4986e2ffe7cf5098e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/2172059
Tested-by: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>