mouse: Use new gesture type for wheel scrolling
This allows us to add fields for wheel ticks to the new details struct,
without cluttering the scroll gesture details. The change is only
enabled when a gesture property is set for now, so that it can be
submitted early without breaking Chromium (avoiding build breakage from
unavoidable skew when the Chromium changes that use this are submitted).

To test, patch https://crrev.com/c/2314657 in Chromium.

BUG=chromium:1077644
TEST=Scroll with a mouse wheel, check movements are correct (including
     very slow scrolls on high-res wheels, e.g. MS Surface Precision).
     Toggle Australian scrolling and check the signs of dy and ticks_dy
     still match.

Change-Id: I3e73d5b7d2e1e40b7795b45a9b434d113624fed3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/2301084
Tested-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>