Remove delay for entering tap-to-drag state
As tap-to-drag is now an opt-in setting, we can be more aggressive when
detecting tap-to-drag gestures.  Before, the user needed to leave their
finger stationary for a short moment before dragging:
  1. Finger down
  2. Finger up
  3. Finger down
  4. Finger remains stationary for 50 ms
  5. Finger moves
This was to prevent false detection, when the user wanted to tap and
then move the cursor.

This CL removes the requirement that the user keep their finger
stationary for a moment. We still differentiate between tap-to-drag
and tap-then-move by the length of the delay between tap and move.

BUG=chromium:840072
TEST=unit tests and touchtests passed.  Manual testing.

Change-Id: Ic7bb99aad3b50581b3cf3457bf69f098623a1329
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/2311355
Tested-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>