Disable zero-finger click and add box filter on by default.
Also, remove an outdated test from Cr-48 that is no longer
applicable. For the Tap To Click state machine test, it doesn't call
into the rest of immediate interpreter to know if the user is
physically clicking the button when zero-finger taps are
disabled. This is a failure in the test, not the production code, and
it's not trivial to fix. To remedy the test, I enable zero-finger
clicks for that test.

BUG=b:176752176
TEST=touchtests shows no serious regressions

Change-Id: I6b7fa523fa9f0c1e666f67b5ec90b08fe96f791f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/2610170
Tested-by: Andrew de los Reyes <adlr@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Andrew de los Reyes <adlr@chromium.org>