Haptic: Use only one finger for force threshold
When adding multiple fingers to a touchpad we add the force of all
fingers to determine if the haptic button is being pressed.  As a
result, it is easy to trigger accidental clicks. Instead of adding
the force of all fingers touching the touchpad, we should use the
maximum force applied by any one finger. This results in a more
intuitive experience.

BUG=b:215423163
TEST=manual testing, update unittests

Change-Id: If2dc1f4ebfe2af368603bccbd6a30b4d143e8f05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3478897
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>