Dynamically adjust force threshold for haptic buttons
When a user presses very hard on a haptic touchpad, this can sometimes
cause hysteresis in the force reports from the touchpad. In order to
avoid a "sticky" feeling touchpad, this CL dynamically adjusts the force
thresholds for physical button click.

After triggering a physical button click, we will increase the threshold
for "button up" based on the maximum force reached while the button is
pressed.

We also dynamically increase the threshold for a subsequent "button
down", but allow that threshold to decrease to normal if we see low
enough force reports.

BUG=b:222170524
TEST=manual, unittests, touchtests

Change-Id: I0e4f9c89f58fae98147345e08e4ab190680f378a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3566340
Tested-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>