Fix callback handling in haptic button generator
When adding a callback to the haptic button generator, we checked
whether the callback for "downstream" interpreters seemed to be
spurious, and didn't forward "spurious" callbacks. However, these
callbacks were not in fact spurious so stopping them early caused issues
with tap-to-click and tap-drag handling in the ImmediateInterpreter.

In the long term, we will refactor callback handling, but for now this
CL forwards all callbacks whether they seem spurious or not.

BUG=b:228095934
TEST=manual testing, unittests, touchtests regression suite

Change-Id: Ib2747ecb4131fab1b12121f9183f8f4d6395c788
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3574394
Tested-by: Sean O'Brien <seobrien@chromium.org>
Auto-Submit: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>