Replace home-made map and set implementations in gestures library
We used X11 in ChromeOS in a way in which X called our code from within
signal handlers. This meant that no mallocs could be used and, hence, we
could not use map and set from standard library.
Since X11 is no longer used in ChromeOS we should be safe to switch to
map and set from standard library.

Modify immediate interpreter unittest as std::vector::reserve() allocates
only the space for the vector elements but it does not construct them.
This in turn caused segmentation fault when using copy assignment operator
(via std::copy()) since it tried to remove sets that were not properly
initialized and contained dirty incoherent data.
Increase the capacity of the vector to the desired one and insert a range
of elements instead of copying them.

BUG=b:200072742
TEST=FEATURES="test" emerge-brya gestures; test manually on Redrix

Change-Id: I5ee121457a9033cfef8af28c2126e3566dd18a0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3732128
Tested-by: Angela Czubak <aczubak@google.com>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Angela Czubak <aczubak@google.com>