Haptic: Always apply movement release suppression
Before this CL, if we were applying force hysteresis
compensation we wouldn't also suppress click release based on
movement speed. This CL makes it so fast movement always makes
it harder to accidentally release a click, even when we are
compensating for force hysteresis.

BUG=b:241629080
TEST=Apply to user feedback reports

Change-Id: Ia69b813375d3d4e6e697dc829344730186115832
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3877538
Tested-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>