Add palm rejection for haptic button click
Use the finger flags set by the Palm Classifying Filter to determine
if a finger is a palm in the Haptic Button Filter. Don't allow palms
to trigger haptic button clicks.

This CL adds a new palm flag which is only applied to touches that
are classified as palms because they are large, and not because
they are stationary near the edge of the touchpad.  This allows
users to click with a dedicated "clicking finger", a relatively
common use case.

This CL also modifies the Lookahead Filter to pass finger flags
(including palm flags) to upstream filters.

BUG=b:242781342, b:222123067
TEST=manual testing, unittests, touchtest regression suite

Change-Id: I6fbf6e15726a3a61af1621f8552215ffcbf5db4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/3988948
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>