Make Gesture::String available to clients
Judging by the description of the commit that added the method [0], this
was originally hidden because it involved making memory allocation,
which was a problem back in the X11 days. Now, we have similar methods
available publicly, and I don't see any reason to keep this one
internal.

[0]: https://chromium.googlesource.com/chromiumos/platform/gestures/+/ca8593f7ee432c5ea061ed41382236de3d65accd

BUG=b:251196347
TEST=cros_workon_make --board=volteer --test chromeos-base/gestures
TEST=use the method from Android input stack

Change-Id: I34f80b646d2bc9ec319f896c7cd11412358a989d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4028658
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Auto-Submit: Harry Cutts <hcutts@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>