libgestures: CY reenable immediate interpreter pinch tests
This test was disabled because the slow pinch was failing to
return a pinch gesture for the final entry but it was doing
what it was supposed to be doing. There was no movement
between the last and the next to last hwstate. So the pinch
gesture should have arrived on the one before and not on
the non-movement hwstate.  I added a kNull to verify the
gesture was in fact NULL for this last entry and moved the
pinch to the one before.

BUG=b:259582075
TEST=USE="coverage" FEATURES="test noclean" emerge-brya chromeos-base/gestures

Change-Id: Ia8e88351121ac0d73cc0769414d7e80a81d917cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4035489
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>