Late delegate setting to avoid init order issues
When a Property is created, initialized in the constructor
initialization list, and uses 'this' as the delegate, there can
be an ordering issue where the delegate gets called that can use the
Property before it is initialized, possibly causing a crash.

The recommended method to initialize a Property of this type is
to create the Property without the delegate and then in the
Constructor body, after the object is fully initialized, set
the delegate.  This will follow the same execution path as setting
the delegate in the initialization list, including calling the
appropriate WasWritten method, but eliminates the possibility
the Property will be used before it is initialized.

BUG=b:253585974
TEST=USE="coverage" FEATURES="test noclean" emerge-brya chromeos-base/gestures

Change-Id: Ia658ea903656d5ad6be638ca0d49b169cfe3127c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4286619
Code-Coverage: Kenneth Albanowski <kenalba@google.com>
Commit-Queue: Kenneth Albanowski <kenalba@google.com>
Reviewed-by: Kenneth Albanowski <kenalba@google.com>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Tested-by: Denis Brockus <dbrockus@chromium.org>