Remove Property constructor delegate
The delegate used for the Property constructor was
most frequently set to 'this' which in many cases
had not been completely initialized at the point
it was passed.  By not passing a delegate into the
Property constructor it will eliminate the
possibility that 'this' was passed before it was
initialized.
Make changes to Property uses to fit this new
mechanism.

BUG=b:253585974
TEST=USE="coverage" FEATURES="test noclean" emerge-brya chromeos-base/gestures

Change-Id: I1751163a3b945a6517e614ee720e7e1a3c49174e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4295788
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Code-Coverage: Harry Cutts <hcutts@chromium.org>
Tested-by: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>