Remove std::optional<std::wrapped_reference>> from .at()
There was a desire to move to use either a pointer or a
normal reference.  The original .at() returned a normal
reference but did not allow for offsets that were out
of bounds.  I went back to this original and used an
abort for the invalid offset.  I added a death test in
util_unittest to verify that was working.

BUG=b:271591258
TEST=USE="coverage" FEATURES="test noclean" emerge-brya chromeos-base/gestures

Change-Id: Ie6a6716f23d5c8516dab7801098e7f4d60b00bee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4357694
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Henry Barnor <hbarnor@chromium.org>
Auto-Submit: Denis Brockus <dbrockus@chromium.org>
Tested-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>