Apply MSC_TIMESTAMP adjustments to timer callbacks
The TimestampFilterInterpreter adjusts the gesture library's internal
idea of time. However, the adjustments were not applied to timer
callbacks. As a result, it sometimes looked to the gesture library that
a callback had come early, causing the callback to be ignored.

This CL applies the time adjustments to callbacks. Because we don't know
exactly when the callback was request, we use the adjustment that moves
the timestamp forward in time the furthest. This should prevent the
gesture library from thinking the callback is too early.

BUG=b:231621332, b:231355042
TEST=Check behavior in unittests, and add regression test

Change-Id: Iaff23dd7cab8f009789758350007e161effb83c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4528259
Code-Coverage: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>