Replace uses of NULL with nullptr
This improves type safety, since nullptr is understood to be a pointer
by the type system, as opposed to NULL which is considered an integer.
This also lets us simplify a lot of null-checking assertions in tests.

This CL was performed by running a few substitions in Neovim:
* %s/NULL/nullptr/g on all source files
* To remove extraneous casts in unit tests:
  * %s/static_cast<[^>]\+>(nullptr)/nullptr/g
  * %s/reinterpret_cast<[^>]\+>(nullptr)/nullptr/g

...followed by a bit of manual clean-up (e.g. for line wrapping and log
messages).

BUG=None
TEST=unit tests

Change-Id: Id081565a909bb64f7d0671734b1fc7cd6b17878b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4584928
Tested-by: Harry Cutts <hcutts@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>