Tap-to-click: support pads that don't report pressure
Normally, the tap-to-click logic requires touches to meet particular
pressure thresholds to be recognized as taps. When used with a touchpad
that doesn't report pressure [0], the pressure values on finger states
are always 0, meaning that tap-to-click didn't work. It appears that not
meeting these thresholds was also causing delays in the start of pointer
motion for moving fingers.

Add a new field to HardwareProperties specifying whether the touchpad
supports pressure reporting, and disable these pressure thresholds if
not. This fixes tap-to-click and makes pointer motion much more
responsive.

[0]: such as the one on the Lenovo P12 Pro's keyboard pack, OPPO Smart
     Touchpad, or the Sony DualShock 4

BUG=b:292051746,b:281907436
TEST=check tap-to-click works (with one to three fingers) on the Lenovo
     P12 Pro keyboard pack and Sony DualShock 4; run unit tests

Change-Id: I25f92da4e9a3ab65fb61eeac3d55b198be23fc1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4794689
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>