Add ability to limit activity event debug logging
Instead of having a master switch to turn on event debug
logging, I believe due to the static limitations of the
buffer that having the ability to reduce to only what
you want to see, as an option, makes sense.  The event
debugging is reduced to
    // Base Event Types
    Gesture, HardwareState, HandleTimer
    // FilterInterpreter Debug Event Types
    Accel, Box, ClickWiggle, FingerMerge, FlingStop,
    HapticButtonGenerator, Iir, IntegratGesture, Logging,
    Lookahead, Metrics, NonLinearity, PalmClassifying,
    Scaling, SensorJump, SplitCorrecting, StationaryWiggle,
    StuckButtonInhibitor, T5R2Correcting, Timestamp,
    TrendClassifying,
    // Interpreter Debug Event Types
    ImmediateInterpreter, MouseInterpreter,
    MultitouchMouseInterpreter,
NOTE: As of this CL, all of the Base Event Type options are
implemented but the other Debug Event Types only exist for Accel
and Timestamp.  Since this is being added as a framework, I
included the others so when/if they are added they will have
been consistently included.

BUG=b:286851905
TEST=USE="coverage" FEATURES="test noclean" emerge-brya chromeos-base/gesture

Change-Id: I620873bcdc5f7533c6aa635c503dd2e37693c057
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4798937
Tested-by: Denis Brockus <dbrockus@chromium.org>
Commit-Queue: Denis Brockus <dbrockus@chromium.org>
Reviewed-by: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>