Fix tests when VCSID is not available.
activity_log_unittest.cc uses VCSID, which is the commit hash of
the git head. This means that the build depends on the entire
repo history. Such dependency will be problematic when ChromeOS
migrates to building with Bazel, as it will lead to excessive
build cache invalidation and rebuilds.

There already is a workaround for building without VCSID
in activity_log.cc and by moving it to the header we make it
possible to test without the VCSID.

BUG=b:300215846
TEST=CQ
TEST=BOARD=amd64-generic bazel test @portage//chromeos-base/gestures:test

Change-Id: I1d4463a68fad9ff4fab35adac45ea0cfc109f292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4861801
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>