HardwareStateBuffer: return a reference from Get
The Get method always returns the address of an index from its states_
buffer. This means that while it could potentially return a pointer that
is outside of the states_ buffer if its logic was incorrect, it can
never actually return null. This means that we can refactor it to return
a reference instead.

Despite this, there are a number of places where we do null checks on
its return value. Those null checks are removed.

BUG=b:245989146
TEST=run all unit tests on an Android build with address sanitization

Change-Id: Iae8ae676d66177fc528342c62882c700408732c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/4865560
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>