Ensure touchpad stack version is included with logs
Use a pointer instead of a temporary variable to hold the touchpad
stack version used to initialize the gesture library. This ensures
that the touchpad stack version is automatically included when logs
are generated. The temporary variable was getting destroyed when it
went of scope resulting in logs not having the information.

BUG=b:283179221
TEST=CQ
TEST=Deployed to Brya device, confirmed in submitted feedback report.

Change-Id: I57e7652891c65e747b2c983d459979f08d00539a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/5015362
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Tested-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>