Rename METADATA.android file to just METADATA
We originally named it this way because ChromeOS also has METADATA files
that look similar but have a slightly different format, so we wanted to
avoid merge conflicts or confusion if the Gestures library acquired one
in the future. However, the Android compliance tooling doesn't
consistently support METADATA.android files, leading to the compliance
linter missing problems with uprev CLs (causing a lot of build breakages
in one instance, see b/296123023). Since it's not a documented feature,
the compliance team aren't inclined to support it unless it gains some
kind of critical mass (b/296989944).

A ChromeOS code search [0] reveals only two instances of repositories
that have METADATA files and aren't third-party imports of some kind, so
it seems unlikely that one would be added to the Gestures library. Even
if it was, I don't think the merge conflicts would be too troublesome,
so let's just rename the file to the Android standard.

[0]: https://source.chromium.org/search?q=f:%5CbMETADATA%5Cb%20-f:%5Esrc%2Faosp%20case:yes%20-f:third_party&ss=chromiumos

Bug: 296989944
Test: Treehugger
Change-Id: Iab4458b147ea2f6364f2e969d9cb0b911fcef078