Makefile: remove -DXLOGGING compiler flag
This was added back in 2011 (https://crrev.com/c/4137) in conjunction
with moving the library to use the X11 xf86MsgVerb logging function
(which avoided uses of malloc, a source of issues on X11 at the time).
We no longer use this logging function, and the library is no longer
used with X11 display servers, so we should be able to remove this now.

BUG=None
TEST=cros_sdk env FEATURES="test" emerge-${BOARD} chromeos-base/gestures

Change-Id: I24bf883ffde2f488181e767cf0bb4e457f48124b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/5297055
Commit-Queue: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>
Auto-Submit: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Sean O'Brien <seobrien@chromium.org>