NonLinearity: don't try to load data if path unset
Previously, if no data file path was set in the gesture property, we'd
try to load a file called "None" (the default property value), and
(hopefully) that would fail. If there happened to be a file called
"None" in the current directory we'd try to load it as non-linearity
correction data. Otherwise, we'd log an error because of the file
loading failure even though we hadn't expected to find a data file.

BUG=b:314743031
TEST=unit tests; add the NonLinearityFilterInterpreter to the Touchpad2
     stack in gestures.cc, add a log line when attempting to load a
     file, and check the line doesn't appear a touchpad is connected

Change-Id: I010086adbf2b6ca69bda32c8c69fc00a239e5a20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/5365863
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Reviewed-by: Sean O'Brien <seobrien@chromium.org>
Tested-by: Harry Cutts <hcutts@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>