Allow finger lock to switch to a new finger
When there are multiple fingers on the touchpad and the user is not
performing a multi-finger gesture, we will "lock on" to a single finger
to follow for motion tracking. Sometimes, we will accidentally lock on
to the wrong finger, such as a thumb that seems to be moving quickly,
but really isn't.

This CL adds a mechanism to change which finger is being tracked after
the initial lock, if another finger is moving fast enough.

BUG=b:358206010
TEST=CQ, manual testing, new touchtest regression tests

Change-Id: I96d69fc392231366f0f06a29db98ffc29d021675
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/gestures/+/5841662
Reviewed-by: Henry Barnor <hbarnor@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Sean O'Brien <seobrien@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>